% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fd_translate}
\alias{fd_translate}
\title{Translate column names of a data frame or lazy tibble using a dictionary}
\usage{
fd_translate(d, dictionary, from = "old", to = "new")
}
\arguments{
\item{d}{A \code{data.frame} or \code{tbl_lazy} object whose column names need to be translated.}

\item{dictionary}{A \code{data.frame} (or tibble) with at least two columns: \code{from} and \code{to}
(default names are "old" and "new"). The \code{from} column contains the current column names
in \code{d}, and the \code{to} column contains the new column names to translate to.}

\item{from}{A string specifying the column name in \code{dictionary} to use for current column
name matching. Defaults to "old".}

\item{to}{A string specifying the column name in \code{dictionary} with new column names
to translate to. Defaults to "new".}
}
\value{
An object of the same class as \code{d} with column names translated.
}
\description{
This function allows renaming column names in an input object (a \code{data.frame} or \code{tbl_lazy})
by translating them with a user-supplied dictionary. Users can choose to translate
column names from "old" to "new" or vice versa.
}
\examples{
# Example dictionary
dictionary <- data.frame(
    old = c("Ship", "SweepLngt"),
    new = c("Platform", "SweepLength")
)
# Example data
df <- data.frame(Ship = "26D4", SweepLngt = 110)

fd_translate(df, dictionary)
}
