\name{OSFD-package}
\alias{OSFD-package}
\docType{package}
\title{
  A sequential algorithm to generate designs that fill the output space
}
\description{
  A sequential method to generate a design that produces points filling the output space. The underlying mapping \code{f} from input space to output space is assumed to be a black-box function that can be evaluated in the forward direction. Please see Wang et al. (2023) for details.
}

\author{
Shangkun Wang, V. Roshan Joseph

Maintainer: Shangkun Wang <sk_wang@gatech.edu>
}
\references{
Wang, Shangkun, Adam P. Generale, Surya R. Kalidindi, and V. Roshan Joseph. "Sequential Designs for Filling Output Spaces." Technometrics, to appear (2023).
}
