\name{OSFD-package}
\alias{OSFD-package}
\docType{package}
\title{
  Sequential algorithms to generate designs that fill the output space.
}
\description{
  Sequential algorithms to generate a design that produces points filling the output space. The underlying mapping \code{f} from input space to output space is assumed to be a black-box function that can be evaluated in the forward direction. Please see Wang et al. (2024) and Wang & Joseph (2025) for details.
}

\author{
Shangkun Wang, V. Roshan Joseph

Maintainer: Shangkun Wang <sk_wang@gatech.edu>
}
\references{
Wang, Shangkun, Adam P. Generale, Surya R. Kalidindi, and V. Roshan Joseph. (2024), "Sequential designs for filling output spaces", Technometrics, 66, 65–76.

Wang, Shangkun, and V. Roshan Joseph. (2025), "Comment: A Model-free Method for Input-Output Space-Filling Design." Technometrics, to appear.
}
